Implemented requested change PR #2
Moved tests/e2e_new to tests/e2e folder
