+++ package.json
@@ -9,7 +9,7 @@
     "postinstall": "sh installers/postinstall/postinstall.sh",
     "test": "NODE_ENV=test ./node_modules/mocha/bin/mocha tests --recursive",
     "test:unit": "NODE_ENV=test ./node_modules/mocha/bin/mocha tests/unit --recursive",
-    "test:e2e": "NODE_ENV=test ./node_modules/mocha/bin/mocha tests/e2e_new --recursive",
+    "test:e2e": "NODE_ENV=test ./node_modules/mocha/bin/mocha tests/e2e --recursive",
     "config:check": "node tests/configs/check_config.js"
   },
   "repository": {
