+++ tests/e2e/modules/calendar_spec.js
@@ -1,19 +1,32 @@
-const globalSetup = require("../global-setup");
-const serverBasicAuth  = require("../../servers/basic-auth.js");
-const app = globalSetup.app;
-const chai = require("chai");
-const expect = chai.expect;
+const helpers = require("../global-setup");
+const path = require("path");
+const request = require("request");
+const serverBasicAuth = require("../../servers/basic-auth.js");
 
-describe("Calendar module", function () {
+const expect = require("chai").expect;
 
-	this.timeout(20000);
+const describe = global.describe;
+const it = global.it;
+const beforeEach = global.beforeEach;
+const afterEach = global.afterEach;
 
-	beforeEach(function (done) {
-		app.start().then(function() { done(); } );
+describe("Calendar module", function() {
+	helpers.setupTimeout(this);
+
+	var app = null;
+
+	beforeEach(function() {
+		return helpers
+			.startApplication({
+				args: ["js/electron.js"]
+			})
+			.then(function(startedApp) {
+				app = startedApp;
+			});
 	});
 
-	afterEach(function (done) {
-		app.stop().then(function() { done(); });
+	afterEach(function() {
+		return helpers.stopApplication(app);
 	});
 
 	describe("Default configuration", function() {
@@ -22,11 +35,10 @@ describe("Calendar module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/calendar/default.js";
 		});
 
-		it("Should return TestEvents", function () {
+		it("Should return TestEvents", function() {
 			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
 		});
 	});
-
 
 	describe("Basic auth", function() {
 		before(function() {
@@ -35,11 +47,14 @@ describe("Calendar module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/calendar/basic-auth.js";
 		});
 
-		it("Should return TestEvents", function () {
+		after(function(done) {
+			serverBasicAuth.close(done());
+		});
+
+		it("Should return TestEvents", function() {
 			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
 		});
 	});
-
 
 	describe("Basic auth by default", function() {
 		before(function() {
@@ -48,7 +63,11 @@ describe("Calendar module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/calendar/auth-default.js";
 		});
 
-		it("Should return TestEvents", function () {
+		after(function(done) {
+			serverBasicAuth.close(done());
+		});
+
+		it("Should return TestEvents", function() {
 			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
 		});
 	});
@@ -60,7 +79,11 @@ describe("Calendar module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/calendar/old-basic-auth.js";
 		});
 
-		it("Should return TestEvents", function () {
+		after(function(done) {
+			serverBasicAuth.close(done());
+		});
+
+		it("Should return TestEvents", function() {
 			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
 		});
 	});
@@ -72,10 +95,12 @@ describe("Calendar module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/calendar/fail-basic-auth.js";
 		});
 
-		it("Should return No upcoming events", function () {
+		after(function(done) {
+			serverBasicAuth.close(done());
+		});
+
+		it("Should return No upcoming events", function() {
 			return app.client.waitUntilTextExists(".calendar", "No upcoming events.", 10000);
 		});
 	});
-
-
 });
