+++ tests/e2e/modules/clock_es_spec.js
@@ -1,8 +1,32 @@
-const globalSetup = require("../global-setup");
-const app = globalSetup.app;
+const helpers = require("../global-setup");
+const path = require("path");
+const request = require("request");
 
-describe("Clock set to spanish language module", function () {
-	this.timeout(20000);
+const expect = require("chai").expect;
+
+const describe = global.describe;
+const it = global.it;
+const beforeEach = global.beforeEach;
+const afterEach = global.afterEach;
+
+describe("Clock set to spanish language module", function() {
+	helpers.setupTimeout(this);
+
+	var app = null;
+
+	beforeEach(function() {
+		return helpers
+			.startApplication({
+				args: ["js/electron.js"]
+			})
+			.then(function(startedApp) {
+				app = startedApp;
+			});
+	});
+
+	afterEach(function() {
+		return helpers.stopApplication(app);
+	});
 
 	describe("with default 24hr clock config", function() {
 		before(function() {
@@ -10,24 +34,14 @@ describe("Clock set to spanish language module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/es/clock_24hr.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
-		it("shows date with correct format", function () {
+		it("shows date with correct format", function() {
 			const dateRegex = /^(?:lunes|martes|miércoles|jueves|viernes|sábado|domingo), \d{1,2} de (?:enero|febrero|marzo|abril|mayo|junio|julio|agosto|septiembre|octubre|noviembre|diciembre) de \d{4}$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .date").should.eventually.match(dateRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .date").should.eventually.match(dateRegex);
 		});
 
 		it("shows time in 24hr format", function() {
-			const timeRegex = /^(?:2[0-3]|[01]\d):[0-5]\d[0-5]\d$/
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .time").should.eventually.match(timeRegex);
+			const timeRegex = /^(?:2[0-3]|[01]\d):[0-5]\d[0-5]\d$/;
+			return app.client.waitUntilWindowLoaded().getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
 
@@ -37,24 +51,14 @@ describe("Clock set to spanish language module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/es/clock_12hr.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
-		it("shows date with correct format", function () {
+		it("shows date with correct format", function() {
 			const dateRegex = /^(?:lunes|martes|miércoles|jueves|viernes|sábado|domingo), \d{1,2} de (?:enero|febrero|marzo|abril|mayo|junio|julio|agosto|septiembre|octubre|noviembre|diciembre) de \d{4}$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .date").should.eventually.match(dateRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .date").should.eventually.match(dateRegex);
 		});
 
 		it("shows time in 12hr format", function() {
 			const timeRegex = /^(?:1[0-2]|[1-9]):[0-5]\d[0-5]\d[ap]m$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .time").should.eventually.match(timeRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
 
@@ -64,18 +68,9 @@ describe("Clock set to spanish language module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/es/clock_showPeriodUpper.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
 		it("shows 12hr time with upper case AM/PM", function() {
 			const timeRegex = /^(?:1[0-2]|[1-9]):[0-5]\d[0-5]\d[AP]M$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .time").should.eventually.match(timeRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
 });
