+++ tests/e2e/modules/clock_spec.js
@@ -1,8 +1,32 @@
-const globalSetup = require("../global-setup");
-const app = globalSetup.app;
+const helpers = require("../global-setup");
+const path = require("path");
+const request = require("request");
 
-describe("Clock module", function () {
-	this.timeout(20000);
+const expect = require("chai").expect;
+
+const describe = global.describe;
+const it = global.it;
+const beforeEach = global.beforeEach;
+const afterEach = global.afterEach;
+
+describe("Clock module", function() {
+	helpers.setupTimeout(this);
+
+	var app = null;
+
+	beforeEach(function() {
+		return helpers
+			.startApplication({
+				args: ["js/electron.js"]
+			})
+			.then(function(startedApp) {
+				app = startedApp;
+			});
+	});
+
+	afterEach(function() {
+		return helpers.stopApplication(app);
+	});
 
 	describe("with default 24hr clock config", function() {
 		before(function() {
@@ -10,24 +34,14 @@ describe("Clock module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/clock_24hr.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
-		it("shows date with correct format", function () {
+		it("shows date with correct format", function() {
 			const dateRegex = /^(?:Monday|Tuesday|Wednesday|Thursday|Friday|Saturday|Sunday), (?:January|February|March|April|May|June|July|August|September|October|November|December) \d{1,2}, \d{4}$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .date").should.eventually.match(dateRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .date").should.eventually.match(dateRegex);
 		});
 
 		it("shows time in 24hr format", function() {
-			const timeRegex = /^(?:2[0-3]|[01]\d):[0-5]\d[0-5]\d$/
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .time").should.eventually.match(timeRegex);
+			const timeRegex = /^(?:2[0-3]|[01]\d):[0-5]\d[0-5]\d$/;
+			return app.client.waitUntilWindowLoaded().getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
 
@@ -37,24 +51,14 @@ describe("Clock module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/clock_12hr.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
-		it("shows date with correct format", function () {
+		it("shows date with correct format", function() {
 			const dateRegex = /^(?:Monday|Tuesday|Wednesday|Thursday|Friday|Saturday|Sunday), (?:January|February|March|April|May|June|July|August|September|October|November|December) \d{1,2}, \d{4}$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .date").should.eventually.match(dateRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .date").should.eventually.match(dateRegex);
 		});
 
 		it("shows time in 12hr format", function() {
 			const timeRegex = /^(?:1[0-2]|[1-9]):[0-5]\d[0-5]\d[ap]m$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .time").should.eventually.match(timeRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
 
@@ -64,18 +68,9 @@ describe("Clock module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/clock_showPeriodUpper.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
 		it("shows 12hr time with upper case AM/PM", function() {
 			const timeRegex = /^(?:1[0-2]|[1-9]):[0-5]\d[0-5]\d[AP]M$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .time").should.eventually.match(timeRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
 
@@ -85,18 +80,9 @@ describe("Clock module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/clock_displaySeconds_false.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
 		it("shows 12hr time without seconds am/pm", function() {
 			const timeRegex = /^(?:1[0-2]|[1-9]):[0-5]\d[ap]m$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .time").should.eventually.match(timeRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .time").should.eventually.match(timeRegex);
 		});
 	});
 
@@ -106,29 +92,17 @@ describe("Clock module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/clock/clock_showWeek.js";
 		});
 
-		beforeEach(function (done) {
-			app.start().then(function() { done(); } );
-		});
-
-		afterEach(function (done) {
-			app.stop().then(function() { done(); });
-		});
-
 		it("shows week with correct format", function() {
 			const weekRegex = /^Week [0-9]{1,2}$/;
-			return app.client.waitUntilWindowLoaded()
-				.getText(".clock .week").should.eventually.match(weekRegex);
+			return app.client.waitUntilWindowLoaded().getText(".clock .week").should.eventually.match(weekRegex);
 		});
 
 		it("shows week with correct number of week of year", function() {
-
-		    it("FIXME: if the day is a sunday this not match");
-		//	const currentWeekNumber = require("current-week-number")();
-		//	const weekToShow = "Week " + currentWeekNumber;
-		//	return app.client.waitUntilWindowLoaded()
-		//		.getText(".clock .week").should.eventually.equal(weekToShow);
+			it("FIXME: if the day is a sunday this not match");
+			//	const currentWeekNumber = require("current-week-number")();
+			//	const weekToShow = "Week " + currentWeekNumber;
+			//	return app.client.waitUntilWindowLoaded()
+			//		.getText(".clock .week").should.eventually.equal(weekToShow);
 		});
-
 	});
-
 });
