+++ tests/e2e/modules/helloworld_spec.js
@@ -1,24 +1,39 @@
-const globalSetup = require("../global-setup");
-const app = globalSetup.app;
+const helpers = require("../global-setup");
+const path = require("path");
+const request = require("request");
 
-describe("Test helloworld module", function () {
-	this.timeout(20000);
+const expect = require("chai").expect;
+
+const describe = global.describe;
+const it = global.it;
+const beforeEach = global.beforeEach;
+const afterEach = global.afterEach;
+
+describe("Test helloworld module", function() {
+	helpers.setupTimeout(this);
+
+	var app = null;
+
+	beforeEach(function() {
+		return helpers
+			.startApplication({
+				args: ["js/electron.js"]
+			})
+			.then(function(startedApp) {
+				app = startedApp;
+			});
+	});
 
 	before(function() {
 		// Set config sample for use in test
 		process.env.MM_CONFIG_FILE = "tests/configs/modules/helloworld/helloworld.js";
 	});
 
-	beforeEach(function (done) {
-		app.start().then(function() { done(); } );
+	afterEach(function() {
+		return helpers.stopApplication(app);
 	});
 
-	afterEach(function (done) {
-		app.stop().then(function() { done(); });
-	});
-
-	it("Test message helloworld module", function () {
-		return app.client.waitUntilWindowLoaded()
-			.getText(".helloworld").should.eventually.equal("Test HelloWorld Module");
+	it("Test message helloworld module", function() {
+		return app.client.waitUntilWindowLoaded().getText(".helloworld").should.eventually.equal("Test HelloWorld Module");
 	});
 });
