+++ tests/e2e/modules/newsfeed_spec.js
@@ -1,27 +1,39 @@
-const globalSetup = require("../global-setup");
-const app = globalSetup.app;
-const chai = require("chai");
-const expect = chai.expect;
+const helpers = require("../global-setup");
+const path = require("path");
+const request = require("request");
 
-describe("Newsfeed module", function () {
+const expect = require("chai").expect;
 
-	this.timeout(20000);
+const describe = global.describe;
+const it = global.it;
+const beforeEach = global.beforeEach;
+const afterEach = global.afterEach;
 
-	beforeEach(function (done) {
-		app.start().then(function() { done(); } );
+describe("Newsfeed module", function() {
+	helpers.setupTimeout(this);
+
+	var app = null;
+
+	beforeEach(function() {
+		return helpers
+			.startApplication({
+				args: ["js/electron.js"]
+			})
+			.then(function(startedApp) {
+				app = startedApp;
+			});
 	});
 
-	afterEach(function (done) {
-		app.stop().then(function() { done(); });
+	afterEach(function() {
+		return helpers.stopApplication(app);
 	});
 
 	describe("Default configuration", function() {
-
 		before(function() {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/newsfeed/default.js";
 		});
 
-		it("show title newsfeed", function () {
+		it("show title newsfeed", function() {
 			return app.client.waitUntilTextExists(".newsfeed .small", "Rodrigo Ramirez Blog", 10000).should.be.fulfilled;
 		});
 	});
