--- tests/e2e_new/modules_position_spec.js
@@ -1,50 +1,0 @@
-const helpers = require("./global-setup");
-const path = require("path");
-const request = require("request");
-
-const expect = require("chai").expect;
-
-const describe = global.describe;
-const it = global.it;
-const beforeEach = global.beforeEach;
-const afterEach = global.afterEach;
-
-describe("Position of modules", function () {
-	helpers.setupTimeout(this);
-
-	var app = null;
-
-	beforeEach(function () {
-		return helpers.startApplication({
-			args: ["js/electron.js"]
-		}).then(function (startedApp) { app = startedApp; })
-	});
-
-	afterEach(function () {
-		return helpers.stopApplication(app);
-	});
-
-	describe("Using helloworld", function () {
-
-		before(function () {
-			// Set config sample for use in test
-			process.env.MM_CONFIG_FILE = "tests/configs/modules/positions.js";
-		});
-
-		var positions = ["top_bar", "top_left", "top_center", "top_right", "upper_third",
-			"middle_center", "lower_third", "bottom_left", "bottom_center", "bottom_right",
-			"bottom_bar", "fullscreen_above", "fullscreen_below"];
-
-		var position;
-		var className;
-		for (idx in positions) {
-			position = positions[idx];
-			className = position.replace("_", ".");
-			it("show text in " + position, function () {
-				return app.client.waitUntilWindowLoaded()
-					.getText("." + className).should.eventually.equal("Text in " + position);
-			});
-		}
-	});
-
-});
